refactoring: moved coordinator into his pkg
