+++ src/main/java/com/example/primo_progetto_spring/Coordinator/repository/CoordinatorRepository.java
@@ -1,6 +1,6 @@
-package com.example.primo_progetto_spring.repository;
+package com.example.primo_progetto_spring.Coordinator.repository;
 
-import com.example.primo_progetto_spring.Entity.Coordinator;
+import com.example.primo_progetto_spring.Coordinator.entity.Coordinator;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
